common/obj_bencher.cc: don't encode/decode 'size_t'
encode/decode funtions for type 'size_t' on OSX are ambiguous

+++ src/common/obj_bencher.cc
@@ -15,6 +15,7 @@
  * try and bench on a pool you don't have permission to access
  * it will just loop forever.
  */
+#include <pthread.h>
 #include "common/Cond.h"
 #include "obj_bencher.h"
 
@@ -206,7 +207,8 @@ void *ObjBencher::status_printer(void *_bencher) {
 
 int ObjBencher::aio_bench(
   int operation, int secondsToRun,
-  int concurrentios, size_t op_size, size_t object_size,
+  int concurrentios,
+  uint64_t op_size, uint64_t object_size,
   unsigned max_objects,
   bool cleanup, const std::string& run_name, bool no_verify) {
 
@@ -222,7 +224,7 @@ int ObjBencher::aio_bench(
 
   //get data from previous write run, if available
   if (operation != OP_WRITE) {
-    size_t prev_op_size, prev_object_size;
+    uint64_t prev_op_size, prev_object_size;
     r = fetch_bench_metadata(run_name_meta, &prev_op_size, &prev_object_size,
 			     &num_objects, &prevPid);
     if (r < 0) {
@@ -331,7 +333,7 @@ static T vec_stddev(vector<T>& v)
 }
 
 int ObjBencher::fetch_bench_metadata(const std::string& metadata_file,
-				     size_t *op_size, size_t* object_size,
+				     uint64_t *op_size, uint64_t* object_size,
 				     int* num_objects, int* prevPid) {
   int r = 0;
   bufferlist object_data;
@@ -1065,7 +1067,7 @@ int ObjBencher::rand_read_bench(int seconds_to_run, int num_objects, int concurr
 
 int ObjBencher::clean_up(const std::string& prefix, int concurrentios, const std::string& run_name) {
   int r = 0;
-  size_t op_size, object_size;
+  uint64_t op_size, object_size;
   int num_objects;
   int prevPid;
